tidying up plotting functions
